refactor: reorganiza estructura de la API en módulos y routers
-Mueve los archivos dentro de carpeta ''app'' - Separa modelos Pydantic en app/models/schemas.py
- Divide endpoints en routers: app/routers/predict.py y app/routers/health.py
- Agrega archivos __init__.py para convertir carpetas en paquetes

closes #11
